+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/config/group/EndpointNameGroupingRuleWatcher.java
@@ -23,6 +23,7 @@ import java.io.StringReader;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.skywalking.oap.server.configuration.api.ConfigChangeWatcher;
 import org.apache.skywalking.oap.server.core.CoreModule;
+import org.apache.skywalking.oap.server.core.config.group.uri.quickmatch.QuickUriGroupingRule;
 import org.apache.skywalking.oap.server.library.module.ModuleProvider;
 import org.apache.skywalking.oap.server.library.util.ResourceUtils;
 
@@ -48,7 +49,7 @@ public class EndpointNameGroupingRuleWatcher extends ConfigChangeWatcher {
     public void notify(final ConfigChangeEvent value) {
         if (value.getEventType().equals(EventType.DELETE)) {
             ruleSetting = null;
-            grouping.setEndpointGroupingRule(new EndpointGroupingRule());
+            grouping.setEndpointGroupingRule(new QuickUriGroupingRule());
         } else {
             ruleSetting = value.getNewValue();
             grouping.setEndpointGroupingRule(new EndpointGroupingRuleReader(new StringReader(ruleSetting)).read());