+++ docs/en/changes/changes.md
@@ -55,6 +55,7 @@
 * Add `ai-pipeline` module.
 * Support HTTP URI formatting through `ai-pipeline` to do pattern recognition.
 * Add new HTTP URI grouping engine with benchmark. 
+* [Breaking Change] Use the new HTTP URI grouping engine to replace the old regex based mechanism.
 
 #### UI
 * Revert: cpm5d function. This feature is cancelled from backend.